Updated Poses and launch logic (League)
