--- Team5898/src/main/java/org/firstinspires/ftc/team5898/PedroPathing_Back.java
@@ -1,332 +1,0 @@
-package org.firstinspires.ftc.team5898;
-
-import com.pedropathing.follower.Follower;
-import com.pedropathing.geometry.BezierCurve;
-import com.pedropathing.geometry.BezierLine;
-import com.pedropathing.geometry.Pose;
-import com.pedropathing.paths.PathChain;
-import com.pedropathing.util.Timer;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.PIDFCoefficients;
-
-import org.firstinspires.ftc.team5898.Constants.CannonConstants;
-import org.firstinspires.ftc.team5898.pedroPathing.Constants;
-
-@Autonomous(name = "PedroPathing Blue - Fixed")
-@Disabled
-public class PedroPathing_Back extends OpMode {
-    private Follower follower;
-    private Timer pathTimer, opModeTimer;
-
-    // Flywheel hardware
-    private DcMotorEx topLauncher, bottomLauncher;
-    private DcMotor frontIntake;
-    private CRServo backLeftServo, backRightServo;
-
-    // Flywheel constants
-    private final double LAUNCH_VELOCITY = CannonConstants.LAUNCH_VELOCITY;
-    private final double INTAKE_POWER = CannonConstants.IntakePower;
-    private final Integer P = CannonConstants.kP;
-    private final Double I = CannonConstants.kI;
-    private final Double D = CannonConstants.kD;
-    private final Integer F = CannonConstants.kF;
-
-    // State flags to prevent repeated calls
-    private boolean launchersStarted = false;
-    private boolean intakeStarted = false;
-    private boolean shootSequenceStarted = false;
-
-    public enum PathState {
-        // START POSITION - END POSITION
-        // DRIVE > MOVEMENT STATE
-        // SHOOT > SCORING
-        DRIVE_STARTPOS_OUTPOS,
-        DRIVE_OUTPOS_PARKPOS,
-        IDLE
-    }
-
-    PathState pathState;
-    private void runShootSequence(String label, PathState nextState){
-        if (follower.isBusy()) {
-            telemetry.addLine(label);
-            return;}
-        if (!shootSequenceStarted){
-            shootSequenceStarted = true;
-            launchersStarted = false;
-            intakeStarted = false;
-
-            pathTimer.resetTimer();
-            startLaunchers();
-            launchersStarted = true;
-
-            telemetry.addLine("Arrived. Spooling for: " + label);
-        }
-
-        double time = pathTimer.getElapsedTimeSeconds();
-        if( time>0.5 && time < 2 && !intakeStarted){
-            backLeftServo.setPower(-.2);
-            backRightServo.setPower(-.2);
-        }
-        if (time > 2.3 && time < 3 && !intakeStarted) {
-            setIntakePower(INTAKE_POWER);
-            intakeStarted = true;
-            telemetry.addLine("Firing sample!");
-        }
-        if (time > 5.25){
-            stopLaunchers();
-            setIntakePower(0);
-            launchersStarted = false;
-            intakeStarted = false;
-            shootSequenceStarted = false;
-            setPathState(nextState);
-        }
-    }
-
-    private final Pose startPose = new Pose(22, 126, Math.toRadians(143));
-    private final Pose shootPose = new Pose(49, 100, Math.toRadians(135));
-    private final Pose PPGPose = new Pose(45, 84, Math.toRadians(180));
-    private final Pose PPGGrabPose = new Pose(16, 84, Math.toRadians(180));
-    private final Pose PGPPose = new Pose(45, 61, Math.toRadians(180));
-    private final Pose PGPGrabPose = new Pose(11, 61, Math.toRadians(180));
-//    private final Pose GPPPose = new Pose(45, 37, Math.toRadians(180));
-//    private final Pose GPPGrabPose = new Pose(13, 37, Math.toRadians(180));
-
-    private final Pose ParkPose = new Pose(33, 75, Math.toRadians(180));
-
-    private PathChain StartToShootPose, ShootToParkPose;
-    private PathChain ShootToPPGPose, GrabPPGPose, PPGToShootPose;
-    private PathChain ShootToPGPPose, GrabPGPPose, PGPToShootPose;
-//    private PathChain ShootToGPPPose, GrabGPPPose, GPPToShootPose;
-
-    public void buildPaths() {
-        StartToShootPose = follower.pathBuilder()
-                .addPath(new BezierLine(startPose, shootPose))
-                .setLinearHeadingInterpolation(startPose.getHeading(), shootPose.getHeading())
-                .build();
-
-        ShootToPPGPose = follower.pathBuilder()
-                .addPath(new BezierLine(shootPose, PPGPose))
-                .setLinearHeadingInterpolation(shootPose.getHeading(), PPGPose.getHeading())
-                .build();
-        GrabPPGPose = follower.pathBuilder()
-                .addPath(new BezierLine(PPGPose, PPGGrabPose))
-                .setLinearHeadingInterpolation(PPGPose.getHeading(), PPGGrabPose.getHeading())
-                .build();
-        PPGToShootPose = follower.pathBuilder()
-                .addPath(new BezierLine(PPGGrabPose, shootPose))
-                .setLinearHeadingInterpolation(PPGGrabPose.getHeading(), shootPose.getHeading())
-                .build();
-
-        ShootToPGPPose = follower.pathBuilder()
-                .addPath(new BezierLine(shootPose, PGPPose))
-                .setLinearHeadingInterpolation(shootPose.getHeading(), PGPPose.getHeading())
-                .build();
-        GrabPGPPose = follower.pathBuilder()
-                .addPath(new BezierLine(PGPPose, PGPGrabPose))
-                .setLinearHeadingInterpolation(PGPPose.getHeading(), PGPGrabPose.getHeading())
-                .build();
-        PGPToShootPose = follower.pathBuilder()
-                .addPath(new BezierCurve(PGPGrabPose,
-                        new Pose(69.00, 61.00),
-                        shootPose))
-                .setLinearHeadingInterpolation(PGPGrabPose.getHeading(), shootPose.getHeading())
-                .build();
-
-//        ShootToGPPPose = follower.pathBuilder()
-//                .addPath(new BezierLine(shootPose, GPPPose))
-//                .setLinearHeadingInterpolation(shootPose.getHeading(), GPPPose.getHeading())
-//                .build();
-//        GrabGPPPose = follower.pathBuilder()
-//                .addPath(new BezierLine(GPPPose, GPPGrabPose))
-//                .setLinearHeadingInterpolation(GPPPose.getHeading(), GPPGrabPose.getHeading())
-//                .build();
-//        GPPToShootPose = follower.pathBuilder()
-//                .addPath(new BezierLine(GPPGrabPose, shootPose))
-//                .setLinearHeadingInterpolation(GPPGrabPose.getHeading(), shootPose.getHeading())
-//                .build();
-
-        ShootToParkPose = follower.pathBuilder()
-                .addPath(new BezierLine(shootPose, ParkPose))
-                .setLinearHeadingInterpolation(shootPose.getHeading(), ParkPose.getHeading())
-                .build();
-    }
-
-    public void statePathUpdate() {
-        switch(pathState) {
-//            case DRIVE_STARTPOS_SHOOTPOS:
-//                follower.followPath(StartToShootPose,true);
-//                setPathState(PathState.SHOOT_PRELOAD);
-//                break;
-//
-//            case SHOOT_PRELOAD:
-//                runShootSequence("PRELOAD", PathState.DRIVE_SHOOTPOS_PPG);
-//                break;
-//
-//            case DRIVE_SHOOTPOS_PPG:
-//                if (!follower.isBusy()) {
-//                    follower.followPath(ShootToPPGPose);
-//                    setPathState(PathState.DRIVE_PPG_GRABPPG);
-//                }
-//                break;
-//
-//            case DRIVE_PPG_GRABPPG:
-//                if (!follower.isBusy()) {
-//                    setIntakePower(INTAKE_POWER);
-//                    follower.followPath(GrabPPGPose);
-//                    setPathState(PathState.DRIVE_GRABPPG_SHOOTPOS);
-//                }
-//                break;
-//
-//            case DRIVE_GRABPPG_SHOOTPOS:
-//                if (!follower.isBusy()) {
-//                    setIntakePower(0);
-//                    follower.followPath(PPGToShootPose,true);
-//                    setPathState(PathState.SHOOT_PPG);
-//                }
-//                break;
-//
-//            case SHOOT_PPG:
-//                runShootSequence("PPG", PathState.DRIVE_SHOOTPOS_PGP);
-//                break;
-//
-//            case DRIVE_SHOOTPOS_PGP:
-//                if (!follower.isBusy()) {
-//                    follower.followPath(ShootToPGPPose);
-//                    setPathState(PathState.DRIVE_PGP_GRABPGP);
-//                }
-//                break;
-//
-//            case DRIVE_PGP_GRABPGP:
-//                if (!follower.isBusy()) {
-//                    setIntakePower(INTAKE_POWER);
-//                    follower.followPath(GrabPGPPose);
-//                    setPathState(PathState.DRIVE_GRABPGP_SHOOTPOS);
-//                }
-//                break;
-//
-//            case DRIVE_GRABPGP_SHOOTPOS:
-//                if (!follower.isBusy()) {
-//                    setIntakePower(0);
-//                    follower.followPath(PGPToShootPose,true);
-//                    setPathState(PathState.SHOOT_PGP);
-//                }
-//                break;
-//
-//            case SHOOT_PGP:
-//                runShootSequence("PGP", PathState.DRIVE_SHOOTPOS_PARKPOS);
-//                break;
-//
-//
-//            case DRIVE_SHOOTPOS_PARKPOS:
-//                if (!follower.isBusy()) {
-//                    follower.followPath(ShootToParkPose);
-//                    setPathState(PathState.IDLE);
-//                }
-//                break;
-
-            case IDLE:
-                telemetry.addLine("Auto complete!");
-                break;
-
-            default:
-                telemetry.addLine("No State Commanded");
-                break;
-        }
-    }
-
-    public void setPathState(PathState newState) {
-        pathState = newState;
-        pathTimer.resetTimer();
-    }
-
-    @Override
-    public void init() {
-        pathState = PathState.DRIVE_STARTPOS_OUTPOS;
-        pathTimer = new Timer();
-        opModeTimer = new Timer();
-        follower = Constants.createFollower(hardwareMap);
-
-        // Initialize flywheel motors
-        topLauncher = hardwareMap.get(DcMotorEx.class, "TLaunch");
-        bottomLauncher = hardwareMap.get(DcMotorEx.class, "BLaunch");
-        topLauncher.setDirection(DcMotorSimple.Direction.REVERSE);
-        bottomLauncher.setDirection(DcMotorSimple.Direction.FORWARD);
-
-        // Set motors to use encoders for velocity control
-        topLauncher.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        bottomLauncher.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        // Set PIDF coefficients for velocity control
-        topLauncher.setPIDFCoefficients(DcMotorEx.RunMode.RUN_USING_ENCODER,
-                new PIDFCoefficients(P, I, D, F));
-        bottomLauncher.setPIDFCoefficients(DcMotorEx.RunMode.RUN_USING_ENCODER,
-                new PIDFCoefficients(P, I, D, F));
-
-        // Set zero power behavior to FLOAT for launchers (reduces resistance)
-        topLauncher.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
-        bottomLauncher.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
-
-        // Initialize intake system
-        frontIntake = hardwareMap.get(DcMotor.class, "Intake");
-        backLeftServo = hardwareMap.get(CRServo.class, "LServo");
-        backRightServo = hardwareMap.get(CRServo.class, "RServo");
-
-        frontIntake.setDirection(DcMotorSimple.Direction.FORWARD);
-        backLeftServo.setDirection(DcMotorSimple.Direction.REVERSE);
-        backRightServo.setDirection(DcMotorSimple.Direction.FORWARD);
-
-        buildPaths();
-        follower.setPose(startPose);
-    }
-
-    @Override
-    public void start() {
-        opModeTimer.resetTimer();
-        setPathState(pathState);
-    }
-
-    /**
-     * Start the flywheel launchers at the configured velocity
-     */
-    private void startLaunchers() {
-        topLauncher.setVelocity(-LAUNCH_VELOCITY);
-        bottomLauncher.setVelocity(-LAUNCH_VELOCITY);
-    }
-
-    /**
-     * Stop the flywheel launchers
-     */
-    private void stopLaunchers() {
-        topLauncher.setPower(0);
-        bottomLauncher.setPower(0);
-    }
-
-    /**
-     * Set power to intake system (front motor and back servos)
-     * @param power Power value for intake system
-     */
-    private void setIntakePower(double power) {
-        frontIntake.setPower(power);
-        backLeftServo.setPower(power);
-        backRightServo.setPower(power);
-    }
-
-    @Override
-    public void loop() {
-        follower.update();
-        statePathUpdate();
-
-        telemetry.addData("Path State", pathState.toString());
-        telemetry.addData("x", follower.getPose().getX());
-        telemetry.addData("y", follower.getPose().getY());
-        telemetry.addData("Heading", follower.getPose().getHeading());
-        telemetry.addData("Path Time", pathTimer.getElapsedTimeSeconds());
-        telemetry.addData("isBusy", follower.isBusy());
-    }
-}
